[IMP] Adjust quant qty calculation and qty_total

+++ product_local_oversea_stock_info/models/product_template.py
@@ -43,7 +43,7 @@ class ProductTemplate(models.Model):
         compute="_compute_qty_local_own_stock",
         store=True,
     )
-    qty_total = fields.Integer(
+    qty_total = fields.Float(
         string="Total Quantity", compute="_compute_qty_total", store=True,
     )
     stock_location = fields.Char(
